Add Gruntfile, package.json and .jshintrc
In order to have a solid build-step and development help, include Grunt.js - for now only with it's default options: linting, concatenation and minification.

There are three tasks available:
* dev - only lints JS
* deploy - linting, concatenation, minifcation
* default - the same as deploy

### Linting
All files that are not third-party-libraries get linted when using

	$ grunt watch

### Concatenation
Concats all files including third-party-libraries into one file in a `dist`-directory.

### Minification
The file generated with `concat` gets minified with Uglify.js

## JSHint options
All options are defaults that should be used in this project:
* Enforcing white-space of 4 spaces per tab
* Using curly braces for block-statements
* Predefined (global) variables